% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestfs.R
\name{nested.glm}
\alias{nested.glm}
\title{Cross-validated generalized linear models}
\usage{
nested.glm(x, y, folds, family = c("binomial", "gaussian"),
  store.glm = FALSE)
}
\arguments{
\item{x}{Dataframe of predictors.}

\item{y}{Outcome variable.}

\item{folds}{Set of cross-validation folds.}

\item{family}{Type of model fitted: \code{"gaussian"} for linear regression
or \code{"binomial"} for logistic regression.}

\item{store.glm}{Whether the object produced by \code{glm} should be
stored.}
}
\value{
A list of length equal to \code{length(folds)}, where each entry contains
the following fields:
\describe{
\item{summary:}{Summary of the fitted model.}
\item{coef:}{Coefficients of the fitted model.}
\item{fit:}{Predicted values for the withdrawn observations.}
\item{obs:}{Observed values for the withdrawn observations.}
\item{test.llk:}{Test log-likelihood.}
\item{test.idx:}{Indices of the the withdrawn observations for this fold.}
\item{regr:}{Object created by glm (only if \code{store.glm=TRUE}).}
}
}
\description{
Run linear or logistic regression on a set of cross-validation folds
}
\details{
This can be used to establish a baseline model, often built only on the
initial set of covariates (those that would be passed through the
\code{init.vars} argument to \code{forward.selection}).
}
\examples{
\dontrun{
   data(diabetes)
   all.folds <- create.folds(10, nrow(diabetes), seed=1)
   base.res <- nested.glm(diabetes[, c("age", "sex", "bmi", "tc",
                                       "ldl", "hdl", "ltg", "glu")],
                          diabetes$Y, all.folds, family="gaussian")
}
}
