\name{nestfs-package}
\alias{nestfs-package}
\alias{nestfs}
\docType{package}
\title{
  Cross-validated (nested) forward selection
}
\description{
  This package provides an implementation of forward selection for the
  based on linear and logistic regression which adopts cross-validation
  as a core component of the selection procedure.
}
\details{
\tabular{ll}{
Package: \tab nestfs\cr
Type: \tab Package\cr
Version: \tab 0.7\cr
Date: \tab 2013-12-21\cr
License: \tab GPL 2.0\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
Marco Colombo

Maintainer: Marco Colombo <m.colombo@ed.ac.uk>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{package}
