% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\alias{X.diab}
\alias{Y.diab}
\title{Diabetes data with interaction terms}
\format{The dataset consists of the following two variables:
\itemize{
  \item X.diab: A dataframe of predictors transformed to have zero mean and
        unit variance.
  \item Y.diab: The outcome variable.
}}
\source{
B. Efron, T. Hastie, I. Johnstone and R. Tibshirani,
"Least angle regression (with discussion)",
\emph{Ann. Statist.}, 32 (2), 407-499, 2004.
\url{http://www.stanford.edu/~hastie/Papers/LARS/data64.txt}
}
\description{
The dataset consists of 442 patients for which a quantitative measure of
disease progression is recorded in the outcome variable \code{Y.diab}. The
dataframe of predictors \code{X.diab} includes 10 baseline measurements, in
addition to 45 interactions and 9 quadratic terms, for a total of 64
variables for each patient.
}
\examples{
data(diabetes, package="nestfs")
}
\keyword{datasets}
