% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes data with interaction terms}
\source{
B. Efron, T. Hastie, I. Johnstone and R. Tibshirani (2004),
Least angle regression, \emph{The Annals of Statistics}, 32 (2), 407-499.
\url{https://doi.org/10.1214/009053604000000067}

The original dataset is available from
\url{https://www.stanford.edu/~hastie/Papers/LARS/data64.txt}
}
\description{
The dataset consists of observations on 442 individuals for which a
quantitative measure of diabetes progression is recorded in variable \code{Y}.
Predictors include 10 baseline measurements, 45 interactions and 9 quadratic
terms, for a total of 64 variables for each individual. All predictors have
been standardized by subtracting the mean and then dividing by the standard
deviation.
}
\examples{
data(diabetes, package="nestfs")
}
\keyword{datasets}
