% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter.predictors}
\alias{filter.predictors}
\title{Filtering of predictors}
\usage{
filter.predictors(x, y, n, ignore = NULL)
}
\arguments{
\item{x}{Design matrix.}

\item{y}{Outcome variable.}

\item{n}{Number of variables to retain.}

\item{ignore}{Names of variables that should be ignored by the filter.}
}
\value{
An array of indices of variables that are retained by the filter.
}
\description{
Filter the predictors, retaining only the top \code{n}.
}
\details{
This performs a univariate test of association of each predictor (not listed
in \code{ignore}) with the outcome, and retains the top \code{n} variables
with smallest p-value according to a Kolmogorov-Smirnov test.
}
\keyword{internal}
