% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestfs.R
\name{nested.forward.selection}
\alias{nested.forward.selection}
\title{Run nested forward selection}
\usage{
nested.forward.selection(x.all, y.all, init.vars, all.folds, ...)
}
\arguments{
\item{x.all}{Dataframe of predictors: this should include all variables in
the initial set and the variables that are allowed to enter the
selected panel.}

\item{y.all}{Outcome variable: if \code{family="binomial"}, it is expected
to only have \code{0-1} entries.}

\item{init.vars}{Initial set of variables (ignored if \code{init.model} is
not \code{NULL}).}

\item{all.folds}{Set of cross-validation folds.}

\item{...}{Arguments to \code{forward.selection}.}
}
\value{
An object of class \code{"nestfs"} of length equal to
\code{length(all.folds)}, where each element is an object of class
\code{"fs"} containing the following additional fields:
\describe{
\item{fit:}{Predicted values for the withdrawn observations.}
\item{obs:}{Observed values for the withdrawn observations.}
\item{test.idx:}{Indices of the the withdrawn observations for this fold.}
\item{model:}{Summary of the model built using the selected panel.}
}
}
\description{
Run nested forward selection starting from a set of variables or a model.
}
\details{
This function allows to obtain an unbiased estimate of the performance
of the selected panels on withdrawn data.
}
\examples{
\dontrun{
data(diabetes)
all.folds <- create.folds(10, nrow(diabetes), seed=1)
nestfs.res <- nested.forward.selection(diabetes[, -1], diabetes$Y,
                                       c("age", "sex"), all.folds,
                                       family=gaussian())
summary(nestfs.res)
}
}
\seealso{
\code{\link{forward.selection}}
}
\keyword{multivariate}
