% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestfs.R
\name{nested.glm}
\alias{nested.glm}
\title{Cross-validated generalized linear models}
\usage{
nested.glm(x, y, folds, family, store.glm = FALSE)
}
\arguments{
\item{x}{Dataframe of predictors.}

\item{y}{Outcome variable. If \code{family=binomial}, it can only contain
two classes of values that can be coerced to 0-1.}

\item{folds}{List of cross-validation folds, where each element contains
the indices of the observations to be withdrawn in that fold.}

\item{family}{Type of model fitted: either \code{gaussian()} for linear
regression or \code{binomial()} for logistic regression. This can be
specified also as a function name (\code{gaussian}) or as a string
(\code{"gaussian"}).}

\item{store.glm}{Whether the object produced by \code{glm} should be
stored (default: \code{FALSE}).}
}
\value{
A list of length equal to \code{length(folds)}, where each entry contains
the following fields:
\describe{
\item{summary:}{Summary of the fitted model.}
\item{coef:}{Coefficients of the fitted model.}
\item{fit:}{Predicted values for the withdrawn observations.}
\item{obs:}{Observed values for the withdrawn observations.}
\item{test.llk:}{Test log-likelihood.}
\item{test.idx:}{Indices of the the withdrawn observations for this fold.}
\item{regr:}{Object created by glm (only if \code{store.glm=TRUE}).}
}
}
\description{
Run linear or logistic regression on a set of cross-validation folds.
}
\details{
This can be used to establish a baseline model, often built only on the
initial set of covariates (those that would be passed through the
\code{init.model} argument to \code{forward.selection}).
}
\examples{
# register a parallel cluster with two cores
library(doParallel)
registerDoParallel(2)

data(diabetes)
folds <- create.folds(10, nrow(X.diab), seed=1)
base.res <- nested.glm(X.diab[, c("age", "sex", "bmi", "tc",
                                  "ldl", "hdl", "ltg", "glu")],
                       Y.diab, folds, gaussian())

# close the parallel cluster
stopImplicitCluster()
}
\keyword{multivariate}
