% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.nestfs}
\alias{print.nestfs}
\alias{print.fs}
\title{Display the contents of (nested) forward selection object}
\usage{
\method{print}{nestfs}(x, ...)

\method{print}{fs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{nestfs} or \code{fs}.}

\item{...}{Further arguments passed to or from other methods.
These are currently ignored.}
}
\value{
The function returns \code{summary} invisibly.
}
\description{
Report summary statistics from a run of (nested) forward selection.
}
\seealso{
\code{\link{forward.selection}} and \code{\link{nested.forward.selection}}.
}
