% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fs}
\alias{summary.fs}
\title{Results summary for forward selection}
\usage{
\method{summary}{fs}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fs}.}

\item{...}{Further arguments passed to or from other methods.
These are currently ignored.}
}
\value{
A dataframe with the following columns:
\describe{
\item{vars:}{Variables in the initial model followed by variables selected.}
\item{fdr:}{False discovery rate, corresponding to the paired test p-values
      computed when the variable was selected.}
\item{llks:}{Validation log-likelihoods.}
\item{diffs:}{Differences in validation log-likelihoods.}
\item{iter:}{Iteration when the variable was selected.}
}
}
\description{
Report summary statistics from a single run of forward selection.
}
\note{
A function of name \code{"getfullname"} to match variable names to full
names is searched on the current workspace, and if found full names are
included in the summary dataframe.
}
